# compiler/prelude: Optimize ASCII string detection.